refactor(app): extract useEngineReload from session-route
UI-triggered engine reload, reload-coordinator registration, the
post-org-onboarding reload latch, server reload-event polling (cursor per
workspace), and desktop engine info move verbatim to
shell/use-engine-reload.ts. Reload events flow through their existing
OpenworkReloadEvent type instead of (item: any). shareWorkspaceState moved
below the hook so routeEngineInfo exists before it is consumed; behavior
unchanged (its outputs are only read later).

The engineInfo() cast is pre-existing (dynamic desktop-bridge Proxy) and
keeps a pointer to the queued typed-IPC work.

session-route.tsx: 2,777 -> 2,676 lines.

Tested: pnpm typecheck clean; bun test 58 pass; pnpm build clean;
pnpm test:e2e 23/23 assertions ok.
